+++ internal/interface/cli/sbi/sbi.go
@@ -22,6 +22,7 @@ func NewSBICommand() *cobra.Command {
 	cmd.AddCommand(NewSBIResetCommand())
 	cmd.AddCommand(NewSBIHistoryCommand())
 	cmd.AddCommand(NewSBIReviewCommand())
+	cmd.AddCommand(NewSBIReportCommand())
 
 	return cmd
 }